Extendable grid

An ExtendableGrid in form of a dictionary with types as keys and type stable value access. This means that grid components are accessed as dict entries, e.g. grid[Coordinates] . The rationale of this approach is explained here.

Notations

A grid is assumed to be a subset of components of a polyhedral complex in d-dimensional space. We distinguish the following element classes characterized by their dimension:

Element class	Meaning

Node	0-dimensional node

Edge	1-dimensional line connecting two neigboring nodes

Face	codimension 1 object separating a cell from outer space or neigboring cell

Cell	codimension 0 object

BFace	Face situated at inner or domain boundary

Region	number to be used to characterize subdomains, contacts etc.

Grid components

Grid components are accessed like Dict entries, the keys must be subtypes of AbstractGridComponent.

Basic set of grid components

Upon construction, an ExtendableGrid needs to be provided with the basic set of grid components denoted by the following component type keys:

Component type key	Meaning

Coordinates	Coordinates of the vertices of the grid cells

CellNodes	Adjacency describing the nodes of grid cell

CellGeometries	Abstract array of subtypes of AbstractElementGeometry describing the geometry of each cell

CellRegions	Abstract array of integers describing region numbers

BFaceNodes	Adjacency structure describing the nodes corresponding to each grid boundary face

BFaceGeometries	Abstract array of subtypes of AbstractElementGeometry describing the geometry of each boundary face

BFaceRegions	Abstract array of integers describig region numbers

CoordinateSystem	Abstract type describing the coordinate system to be used

Hierarchy of component type keys

The list of components can be printed using the gridcomponents method.

AbstractGridComponent
├─ AbstractElementGeometries
│  ├─ BEdgeGeometries
│  ├─ BFaceGeometries
│  ├─ CellGeometries
│  ├─ EdgeGeometries
│  ├─ FaceGeometries
│  ├─ UniqueBEdgeGeometries
│  ├─ UniqueBFaceGeometries
│  ├─ UniqueCellGeometries
│  ├─ UniqueEdgeGeometries
│  └─ UniqueFaceGeometries
├─ AbstractElementRegions
│  ├─ BEdgeRegions
│  ├─ BFaceRegions
│  ├─ CellRegions
│  ├─ EdgeRegions
│  └─ FaceRegions
├─ AbstractGridAdjacency
│  ├─ BEdgeNodes
│  ├─ BFaceCells
│  ├─ BFaceEdges
│  ├─ BFaceNodes
│  ├─ CellEdgeSigns
│  ├─ CellEdges
│  ├─ CellFaceOrientations
│  ├─ CellFaceSigns
│  ├─ CellFaces
│  ├─ CellNodes
│  ├─ EdgeCells
│  ├─ EdgeNodes
│  ├─ FaceCells
│  ├─ FaceEdges
│  └─ FaceNodes
├─ AbstractGridFloatArray1D
│  ├─ BEdgeVolumes
│  ├─ BFaceVolumes
│  ├─ CellVolumes
│  ├─ EdgeVolumes
│  ├─ FaceVolumes
│  ├─ XCoordinates
│  ├─ YCoordinates
│  └─ ZCoordinates
├─ AbstractGridFloatArray2D
│  ├─ Coordinates
│  ├─ EdgeTangents
│  ├─ FaceNormals
│  └─ VoronoiFaceCenters
├─ AbstractGridIntegerArray1D
│  ├─ BEdgeEdges
│  ├─ BFaceCellPos
│  ├─ BFaceFaces
│  ├─ CellParents
│  ├─ NodeInParent
│  └─ NodePatchGroups
├─ AbstractGridIntegerArray2D
├─ BFaceNormals
├─ CoordinateSystem
├─ AbstractGridFloatConstant
├─ AbstractGridIntegerConstant
│  ├─ NumBEdgeRegions
│  ├─ NumBFaceRegions
│  └─ NumCellRegions
└─ ParentGrid

Additional components

Additional components can be added by defining a subtype of AbstractGridComponent or a fitting subtype thereof, and assigning the value to the corresponding Dict entry:

g=simplexgrid([1,2,3,4.0])
abstract type MyComponent <: AbstractGridComponent end
g[MyComponent]=13
show(g)

ExtendableGrid{Float64, Int32};
dim: 1 nodes: 4 cells: 3 bfaces: 2

Alternatively, component creation can be perfomed lazily. For this purpose one needs to define an instantiate method:

abstract type NodeCells <: AbstractGridAdjacency end
ExtendableGrids.instantiate(grid, ::Type{NodeCells})=atranspose(grid[CellNodes])
g=simplexgrid([1,2,3,4.0])
show(g[NodeCells])

1
1 2
2 3
3

Grid API

ExtendableGrids.ElementInfo — Type

const ElementInfo{T}=Union{Vector{T},VectorOfConstants{T}}`

Union type for element information arrays. If all elements have the same information, it can be stored in an economical form as a VectorOfConstants.

ExtendableGrids.AbstractElementGeometries — Type

abstract type AbstractElementGeometries <: AbstractGridComponent

Array of element geometry information.

ExtendableGrids.AbstractElementRegions — Type

abstract type AbstractElementRegions <: AbstractGridComponent

Array of element region number information.

ExtendableGrids.AbstractGridAdjacency — Type

abstract type AbstractGridAdjacency <: AbstractGridComponent

Any kind of adjacency between grid components

ExtendableGrids.AbstractGridComponent — Type

abstract type AbstractGridComponent <: AbstractExtendableGridApexType

Apex type for grid components.

ExtendableGrids.AbstractGridFloatArray1D — Type

abstract type AbstractGridFloatArray1D <: AbstractGridComponent

1D Array of floating point data

ExtendableGrids.AbstractGridFloatArray2D — Type

abstract type AbstractGridFloatArray2D <: AbstractGridComponent

2D Array of floating point data

ExtendableGrids.AbstractGridFloatConstant — Type

abstract type AbstractGridFloatConstant <: AbstractGridComponent

Floating point number

ExtendableGrids.AbstractGridIntegerArray1D — Type

abstract type AbstractGridIntegerArray1D <: AbstractGridComponent

1D Array of interger data

ExtendableGrids.AbstractGridIntegerArray2D — Type

abstract type AbstractGridIntegerArray2D <: AbstractGridComponent

2D Array of integer data

ExtendableGrids.AbstractGridIntegerConstant — Type

abstract type AbstractGridIntegerConstant <: AbstractGridComponent

Integer number

ExtendableGrids.BEdgeRegions — Type

abstract type BEdgeRegions <: AbstractElementRegions

Boundary edge region number per boundary edge

ExtendableGrids.BFaceGeometries — Type

Description of boundary face geometries

abstract type BFaceGeometries <: AbstractElementGeometries

ExtendableGrids.BFaceNodes — Type

abstract type BFaceNodes <: AbstractGridAdjacency

Adjacency describing nodes per grid boundary face

ExtendableGrids.BFaceRegions — Type

abstract type BFaceRegions <: AbstractElementRegions

Boundary region number per boundary face

ExtendableGrids.CellGeometries — Type

abstract type CellGeometries <: AbstractElementGeometries

Description of cell geometries

ExtendableGrids.CellNodes — Type

abstract type CellNodes <: AbstractGridAdjacency

Adjacency describing nodes per grid cell

ExtendableGrids.CellRegions — Type

abstract type CellRegions <: AbstractElementRegions

Cell region number per cell

ExtendableGrids.CoordinateSystem — Type

abstract type CoordinateSystem <: AbstractGridComponent

Coordinate system

ExtendableGrids.Coordinates — Type

abstract type Coordinates <: AbstractGridFloatArray2D

Node coordinates

ExtendableGrids.ExtendableGrid — Type

mutable struct ExtendableGrid{Tc, Ti}

Grid type wrapping Dict

ExtendableGrids.NumBEdgeRegions — Type

abstract type NumBEdgeRegions <: ExtendableGrids.AbstractGridIntegerConstant

Number of boundary edge regions

ExtendableGrids.NumBFaceRegions — Type

abstract type NumBFaceRegions <: ExtendableGrids.AbstractGridIntegerConstant

Number of boundary face regions

ExtendableGrids.NumCellRegions — Type

abstract type NumCellRegions <: ExtendableGrids.AbstractGridIntegerConstant

Number of cell regions

Base.delete! — Method

delete!(grid::ExtendableGrid, T::Type{var"#s5"} where var"#s5"<:AbstractGridComponent) -> Dict{Type{var"#s5"} where var"#s5"<:AbstractGridComponent, Any}

Remove grid component

Base.get! — Method

get!(grid::ExtendableGrid, T::Type{var"#s5"} where var"#s5"<:AbstractGridComponent) -> Any

To be called by getindex. This triggers lazy creation of non-existing gridcomponents

Base.getindex — Method

Base.getindex(grid::ExtendableGrid,T::Type{<:AbstractGridComponent})

Generic method for obtaining grid component.

This method is mutating in the sense that non-existing grid components are created on demand.

Due to the fact that components are stored as Any the return value triggers type instability.

Base.getindex — Method

Base.getindex(grid::ExtendableGrid{Tc,Ti},T::Type{<:AbstractElementGeometries})::ElementInfo{DataType}

Type stable method to obtain element type

Base.getindex — Method

Base.getindex(grid::ExtendableGrid{Tc,Ti},T::Type{<:AbstractElementRegions})::ElementInfo{Ti} where{Tc,Ti}

Type stable method to obtain element region number

Base.getindex — Method

Base.getindex(grid::ExtendableGrid{Tc,Ti},T::Type{<:AbstractGridAdjacency})::Adjacency{Ti} where{Tc,Ti}

Type stable return of adjacency component

Base.getindex — Method

Base.getindex(grid::ExtendableGrid{Tc,Ti},T::Type{<:AbstractGridFloatArray1D})::Array{Tc,1} where{Tc,Ti}

Type stable method to obtain 1D array from grid

Base.getindex — Method

Base.getindex(grid::ExtendableGrid{Tc,Ti},T::Type{<:AbstractGridFloatArray2D})::Array{Tc,2} 

Type stable c method to obtain 2D array from grid

Base.getindex — Method

Base.getindex(grid::ExtendableGrid{Tc,Ti},T::Type{<:AbstractGridFloatConstant})::Tc where{Tc,Ti}

Type stable method to obtain float constant stored on elements

Base.getindex — Method

Base.getindex(grid::ExtendableGrid{Tc,Ti},T::Type{<:AbstractGridIntegerConstant})::Ti where{Tc,Ti}

Type stable method to obtain integer constant stored on elements

Base.haskey — Method

haskey(g::ExtendableGrid, k) -> Bool

Check if key is in grid

Base.keys — Method

keys(g::ExtendableGrid) -> Base.KeySet{Type{var"#s5"} where var"#s5"<:AbstractGridComponent, Dict{Type{var"#s5"} where var"#s5"<:AbstractGridComponent, Any}}

Keys in grid

Base.map — Method

map(f::Function, grid::ExtendableGrid) -> Any

Map a function onto node coordinates of grid

Base.setindex! — Method

setindex!(grid::ExtendableGrid, v, T::Type{var"#s5"} where var"#s5"<:AbstractGridComponent) -> Any

Set new grid component

ExtendableGrids.coord_type — Method

Type of coordinates in grid

ExtendableGrids.dim_grid — Method

dim_grid(grid)

Grid dimension dimension of grid (larges element dimension)

ExtendableGrids.dim_space — Method

dim_space(grid)

Space dimension of grid

ExtendableGrids.gridcomponents — Method

gridcomponents()

Print the hierarchy of grid component key types (subtypes of AbstractGridComponent. This includes additionally user defined subptypes.

ExtendableGrids.index_type — Method

Type of indices

ExtendableGrids.instantiate — Function

"Hook" for methods instantiating lazy components.

ExtendableGrids.instantiate — Method

instantiate(grid, _::Type{NumBEdgeRegions}) -> Any

Instantiate number of boundary edge regions

ExtendableGrids.instantiate — Method

instantiate(grid, _::Type{NumBFaceRegions}) -> Any

Instantiate number of bface regions

ExtendableGrids.instantiate — Method

instantiate(grid, _::Type{NumCellRegions}) -> Any

Instantiate number of cell regions

ExtendableGrids.num_bedgeregions — Method

num_bedgeregions(grid::ExtendableGrid) -> Any

Maximum boundary edge region numbers

ExtendableGrids.num_bedges — Method

num_bedges(grid::ExtendableGrid) -> Int64

Number of boundary edges in grid.

ExtendableGrids.num_bfaceregions — Method

num_bfaceregions(grid::ExtendableGrid) -> Any

Maximum boundary face region numbers

ExtendableGrids.num_bfaces — Method

num_bfaces(grid::ExtendableGrid) -> Int64

Number of boundary faces in grid.

ExtendableGrids.num_cellregions — Method

num_cellregions(grid::ExtendableGrid) -> Any

Maximum cell region number

ExtendableGrids.num_cells — Method

num_cells(grid::ExtendableGrid) -> Int64

Number of cells in grid

ExtendableGrids.num_nodes — Method

num_nodes(grid)

Number of nodes in grid

ExtendableGrids.seemingly_equal — Method

seemingly_equal(array1, array2)

Check for seeming equality of two arrays by random sample.

ExtendableGrids.seemingly_equal — Method

seemingly_equal(grid1, grid2)

Recursively check seeming equality of two grids. Seemingly means that long arrays are only compared via random samples

ExtendableGrids.veryform — Method

veryform(grid::ExtendableGrid, v, _::Type{var"#s5"} where var"#s5"<:AbstractGridComponent) -> Any

Default veryform method.

"veryform" means "verify and/or transform" and is called to check and possibly transform components to be added to the grid via setindex!.

The default method just passes data through.

ExtendableGrids.veryform — Method

veryform(grid::ExtendableGrid{Tc,Ti},v,T::Type{<:AbstractGridAdjacency}) where{Tc,Ti}

Check proper type of adjacencies upon insertion