Derived adjacencies

API

ExtendableGrids.BEdgeAssemblyGroups — Type

abstract type BEdgeAssemblyGroups <: AbstractGridAdjacency

Adjacency describing assembly groups for boundary edges

ExtendableGrids.BEdgeEdges — Type

abstract type BEdgeEdges <: AbstractGridIntegerArray1D

Vector with edge numbers of boundary edges

ExtendableGrids.BEdgeGeometries — Type

abstract type BEdgeGeometries <: AbstractElementGeometries

Vector with boundary edge geometries

ExtendableGrids.BEdgeNodes — Type

abstract type BEdgeNodes <: AbstractGridAdjacency

Adjacency describing nodes for boundary edges

ExtendableGrids.BEdgeVolumes — Type

abstract type BEdgeVolumes <: AbstractGridFloatArray1D

Vector with boundary edge volumes

ExtendableGrids.BFaceAssemblyGroups — Type

abstract type BFaceAssemblyGroups <: AbstractGridAdjacency

Adjacency describing assembly groups for bfaces

ExtendableGrids.BFaceCellPos — Type

abstract type BFaceCellPos <: AbstractGridIntegerArray1D

Vector with local positions of bfaces in their adjacent cell

ExtendableGrids.BFaceFaces — Type

abstract type BFaceFaces <: AbstractGridIntegerArray1D

Vector with face numbers of boundary faces

ExtendableGrids.BFaceVolumes — Type

abstract type BFaceVolumes <: AbstractGridFloatArray1D

Vector with boundary face volumes

ExtendableGrids.CellAssemblyGroups — Type

abstract type CellAssemblyGroups <: AbstractGridAdjacency

Adjacency describing assembly groups for cells

ExtendableGrids.CellEdgeSigns — Type

abstract type CellEdgeSigns <: AbstractGridAdjacency

Adjacency describing edge signs for cells

ExtendableGrids.CellEdges — Type

abstract type CellEdges <: AbstractGridAdjacency

Adjacency describing edges for cells

ExtendableGrids.CellFaceOrientations — Type

abstract type CellFaceOrientations <: AbstractGridAdjacency

Adjacency describing face orientations for cells

ExtendableGrids.CellFaceSigns — Type

abstract type CellFaceSigns <: AbstractGridAdjacency

Adjacency describing face signs for cells

ExtendableGrids.CellFaces — Type

abstract type CellFaces <: AbstractGridAdjacency

Adjacency describing faces for cells

ExtendableGrids.CellVolumes — Type

abstract type CellVolumes <: AbstractGridFloatArray1D

Vector of cell volumes

ExtendableGrids.EdgeAssemblyGroups — Type

abstract type EdgeAssemblyGroups <: AbstractGridAdjacency

Adjacency describing assembly groups for edges

ExtendableGrids.EdgeCells — Type

abstract type EdgeCells <: AbstractGridAdjacency

Adjacency describing cells for edges

ExtendableGrids.EdgeGeometries — Type

abstract type EdgeGeometries <: AbstractElementGeometries

Vector with edge geometries

ExtendableGrids.EdgeNodes — Type

abstract type EdgeNodes <: AbstractGridAdjacency

Adjacency describing nodes for edges

ExtendableGrids.EdgeRegions — Type

abstract type EdgeRegions <: AbstractElementRegions

Vector with edge regions

ExtendableGrids.EdgeTangents — Type

abstract type EdgeTangents <: AbstractGridFloatArray2D

Matrix with edge tangents

ExtendableGrids.EdgeVolumes — Type

abstract type EdgeVolumes <: AbstractGridFloatArray1D

Vector with edge volumes

ExtendableGrids.FaceAssemblyGroups — Type

abstract type FaceAssemblyGroups <: AbstractGridAdjacency

Adjacency describing assembly groups for faces

ExtendableGrids.FaceCells — Type

abstract type FaceCells <: AbstractGridAdjacency

Adjacency describing cells for faces

ExtendableGrids.FaceEdgeSigns — Type

abstract type FaceEdgeSigns <: AbstractGridAdjacency

Adjacency describing edges signs for faces

ExtendableGrids.FaceEdges — Type

abstract type FaceEdges <: AbstractGridAdjacency

Adjacency describing edges for faces

ExtendableGrids.FaceGeometries — Type

abstract type FaceGeometries <: AbstractElementGeometries

Vector of face geometries

ExtendableGrids.FaceNodes — Type

abstract type FaceNodes <: AbstractGridAdjacency

Adjacency describing nodes for faces

ExtendableGrids.FaceNormals — Type

abstract type FaceNormals <: AbstractGridFloatArray2D

Matrix of face normals

ExtendableGrids.FaceRegions — Type

abstract type FaceRegions <: AbstractElementRegions

Vector of face regions

ExtendableGrids.FaceVolumes — Type

abstract type FaceVolumes <: AbstractGridFloatArray1D

Vector of face volumes

ExtendableGrids.NodeCells — Type

abstract type NodeCells <: AbstractGridAdjacency

Adjacency describing cells for nodes (transposed of CellNodes)

ExtendableGrids.NodeEdges — Type

abstract type NodeEdges <: AbstractGridAdjacency

Adjacency describing edges for nodes (transposed of EdgeNodes)

ExtendableGrids.NodeFaces — Type

abstract type NodeFaces <: AbstractGridAdjacency

Adjacency describing faces for nodes (transposed of FaceNodes)

ExtendableGrids.NodePatchGroups — Type

abstract type NodePatchGroups <: AbstractGridIntegerArray1D

Vector with patch groups for nodes (ensuring that node patches of nodes in same group do not overlap)

ExtendableGrids.UniqueBEdgeGeometries — Type

abstract type UniqueBEdgeGeometries <: AbstractElementGeometries

Vector with unique boundary edge geometries

ExtendableGrids.UniqueBFaceGeometries — Type

abstract type UniqueBFaceGeometries <: AbstractElementGeometries

Vector with unique bface geometries

ExtendableGrids.UniqueCellGeometries — Type

abstract type UniqueCellGeometries <: AbstractElementGeometries

Vector of unique cell geometries

ExtendableGrids.UniqueEdgeGeometries — Type

abstract type UniqueEdgeGeometries <: AbstractElementGeometries

Vector with unique edge geometries

ExtendableGrids.UniqueFaceGeometries — Type

abstract type UniqueFaceGeometries <: AbstractElementGeometries

Vector of unique face geometries

ExtendableGrids.BFaceCells — Type

abstract type BFaceCells <: AbstractGridAdjacency

Adjacency describing cells per boundary or interior face

ExtendableGrids.BFaceEdges — Type

abstract type BFaceEdges <: AbstractGridAdjacency

Adjacency describing edges per boundary or interior face

ExtendableGrids.BFaceNormals — Type

abstract type BFaceNormals <: AbstractGridComponent

Adjacency describing outer normals to boundary faces

ExtendableGrids.prepare_edges! — Method

prepare_edges!(grid)

Prepare edge adjacencies (celledges, edgecells, edgenodes)

Currently depends on ExtendableSparse, we may want to remove this adjacency.