Grid partitioning

Access to partitioning data

API calls

ExtendableGrids.num_pcolors — Function

num_pcolors(grid)

Return number of partition colors.

ExtendableGrids.num_partitions — Function

num_partitions(grid)

Return number of partitions.

ExtendableGrids.num_partitions_per_color — Function

num_partitions_per_color(grid)

Return a vector containing the number of partitions for each of the colors of the grid partitioning.

ExtendableGrids.pcolors — Function

pcolors(grid)

Return range of all pcolors.

ExtendableGrids.pcolor_partitions — Function

pcolor_partitions(grid, color)

Return range of partitions for given pcolor.

ExtendableGrids.partition_cells — Function

partition_cells(grid, part)

Return range of cells belonging to a given partition.

ExtendableGrids.partition_nodes — Function

partition_nodes(grid, part)

Return range of nodes belonging to a given partition.

ExtendableGrids.check_partitioning — Function

check_partitioning(grid; verbose)

Check correctness of partittioning

Key types for grid access

ExtendableGrids.PColorPartitions — Type

abstract type PColorPartitions <: AbstractGridIntegerArray1D

Key type describing colors of partitions. These correspond to a coloring of the neigborhood graphs of partitions such that operations (e.g. FEM assembly) on partitions of a given color can be performed in parallel.

grid[PColorPartitions] returns an integer vector describing the partition colors ("pcolors") of a grid. Let p=grid[PColorPartitions]. Then all partitions with numbers i ∈ p[c]:p[c+1]-1 have "color" c.

ExtendableGrids.PartitionCells — Type

abstract type PartitionCells <: AbstractGridIntegerArray1D

Key type describing the cells of a given partition.

grid[PartitionCells] returns an integer vector describing the cells of a partition given by its number. Let pc=grid[PartitionCells]. Then all cells with index i ∈ pc[p]:pc[p+1]-1 belong to partition p.

ExtendableGrids.PartitionNodes — Type

abstract type PartitionNodes <: AbstractGridIntegerArray1D

Key type describing the nodes of a given partition.

grid[PartitionNodes] returns an integer vector describing the nodes of a partition given by its number. Let pn=grid[PartitionNodes]. Then all nodes with index i ∈ pn[p]:pn[p+1]-1 belong to partition p.

ExtendableGrids.NodePermutation — Type

abstract type NodePermutation <: AbstractGridIntegerArray1D

Key type describing the permutation of the nodes of a partitioned grid with respect to the unpartitioned origin.

If pgrid is the partitioned grid and grid is the unpartitioned origin, then

pgrid[Coordinates][:,pgrid[NodePermutation]]==grid[Coordinates]

Partitioning algorithms

ExtendableGrids.AbstractPartitioningAlgorithm — Type

abstract type AbstractPartitioningAlgorithm

Abstract super type for partitioning algorithms

ExtendableGrids.partition — Function

partition(grid, alg)

Partition grid according to alg, such that the neigborhood graph of partitions is colored in such a way, that all partitions with a given color can be worked on in parallel.

ExtendableGrids.TrivialPartitioning — Type

struct TrivialPartitioning <: AbstractPartitioningAlgorithm

Trivial partitioning: all grid cells belong to single partition number 1.

ExtendableGrids.PlainMetisPartitioning — Type

struct PlainMetisPartitioning <: AbstractPartitioningAlgorithm

Subdivide grid into npart partitions using Metis.partition and color the resulting partition neigborhood graph. This requires to import Metis.jl in order to trigger the corresponding extension.

npart::Int64

Internal API

ExtendableGrids.trivial_partitioning! — Function

trivial_partitioning!(grid)

(internal) Create trivial partitioning: the whole grid is partition #1 with just one color.

ExtendableGrids.instantiate — Method

instantiate(grid::ExtendableGrid, ::Type{PColorPartitions})

If not given otherwise, instantiate partition data with trivial partitioning.

ExtendableGrids.instantiate — Method

instantiate(grid::ExtendableGrid, ::Type{PColorPartitions})

If not given otherwise, instantiate partition data with trivial partitioning.

ExtendableGrids.partgraph — Function

partgraph(cellpartitions, ncellpartitions, cellcelladj)

(internal) Create neigbourhood graph for given partitioning.

ExtendableGrids.induce_node_partitioning! — Function

induce_node_partitioning!(grid; keep_nodepermutation)

(internal) Induce node partitioning from partitioning of grid.