Plotting

Provide plotting facilities for various Julia graphics APIs.

API

ExtendableGrids.SubPlotContext — Type

A SubPlotContext is just a dictionary which contains plotting information, including type of the plotter and its position in the plot.

ExtendableGrids.MakieType — Type

abstract type MakieType

Abstract type for dispatching on plotter

ExtendableGrids.PlotContext — Type

struct PlotContext

Context type for plots.

ExtendableGrids.PlotContext — Method

PlotContext(; Plotter, kwargs...)

Create a plot context.

Plotter: defaults to nothing and can be PyPlot, Plots, VTKView, Makie. This pattern to pass the backend as a module to a plot function allows to circumvent to create heavy default package dependencies.

Depending on the layout keyword argument, a 2D grid of subplots is created. Further plot! commands then plot into one of these subplots:

p=PlotContext(Plotter=PyPlot, layout=(2,2)
plot!(p[1,2], ...)
````

A `plot`  command just implicitely creates a plot context:

julia plot(..., Plotter=PyPlot)

is equivalent to

julia p=PlotContext(Plotter=PyPlot, layout=(1,1) plot!(p,...) ```

Please not that the return values of all plot commands are specific to the Plotter.

Depending on the backend, interactive mode switch between "gallery view" showing all plots at onece and "focused view" showing only one plot is possible.

Keyword arguments:

alpha: Surface alpha value. Default: 1.0

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap. Default: viridis

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

elevation: Height factor for elevation of 2D plot. Default: 0.0

show: Show plot immediately. Default: true

legend_location: Location of legend. Default: upper right

outline: Plot outline of domain. Default: true

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

label: Label of plot. Default: f

interior: Plot interior of grid. Default: true

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

title: Plot title. Default:

resolution: Plot xy resolution. Default: (500, 500)

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ExtendableGrids.PlotsType — Type

abstract type PlotsType

Abstract type for dispatching on plotter

ExtendableGrids.PyPlotType — Type

abstract type PyPlotType

Abstract type for dispatching on plotter

ExtendableGrids.VTKViewType — Type

abstract type VTKViewType

Abstract type for dispatching on plotter

Base.getindex — Method

getindex(p, i, j)

Return a SubPlotContext

Base.size — Method

size(p)

Return the layout of a PlotContext

ExtendableGrids.ismakie — Method

ismakie(Plotter)

Heuristically check if Plotter is Makie/WGLMakie

ExtendableGrids.isplots — Method

isplots(Plotter)

Heuristically check if Plotter is Plots

ExtendableGrids.ispyplot — Method

ispyplot(Plotter)

Heuristically check if Plotter is PyPlot

ExtendableGrids.isvtkview — Method

isvtkview(Plotter)

Heuristically check if Plotter is VTKView

ExtendableGrids.plot! — Method

plot!(p, grid, func; kwargs...)

Plot vector on grid as P1 FEM function.

Keyword arguments

alpha: Surface alpha value. Default: 1.0

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap. Default: viridis

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

elevation: Height factor for elevation of 2D plot. Default: 0.0

show: Show plot immediately. Default: true

legend_location: Location of legend. Default: upper right

outline: Plot outline of domain. Default: true

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

label: Label of plot. Default: f

interior: Plot interior of grid. Default: true

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

title: Plot title. Default:

resolution: Plot xy resolution. Default: (500, 500)

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ExtendableGrids.plot! — Method

plot!(p, grid; kwargs...)

Plot grid.

Keyword arguments:

alpha: Surface alpha value. Default: 1.0

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap. Default: viridis

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

elevation: Height factor for elevation of 2D plot. Default: 0.0

show: Show plot immediately. Default: true

legend_location: Location of legend. Default: upper right

outline: Plot outline of domain. Default: true

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

label: Label of plot. Default: f

interior: Plot interior of grid. Default: true

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

title: Plot title. Default:

resolution: Plot xy resolution. Default: (500, 500)

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ExtendableGrids.plot! — Method

plot!(ctx, grid, func; kwargs...)

Plot vector on grid as P1 FEM function.

Keyword arguments

alpha: Surface alpha value. Default: 1.0

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap. Default: viridis

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

elevation: Height factor for elevation of 2D plot. Default: 0.0

show: Show plot immediately. Default: true

legend_location: Location of legend. Default: upper right

outline: Plot outline of domain. Default: true

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

label: Label of plot. Default: f

interior: Plot interior of grid. Default: true

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

title: Plot title. Default:

resolution: Plot xy resolution. Default: (500, 500)

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ExtendableGrids.plot! — Method

plot!(ctx, grid; kwargs...)

Plot grid.

Keyword arguments:

alpha: Surface alpha value. Default: 1.0

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap. Default: viridis

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

elevation: Height factor for elevation of 2D plot. Default: 0.0

show: Show plot immediately. Default: true

legend_location: Location of legend. Default: upper right

outline: Plot outline of domain. Default: true

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

label: Label of plot. Default: f

interior: Plot interior of grid. Default: true

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

title: Plot title. Default:

resolution: Plot xy resolution. Default: (500, 500)

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ExtendableGrids.plot — Method

plot(grid, func; Plotter, kwargs...)

Plot vector on grid without predefined context as P1 FEM function.

Keyword arguments:

alpha: Surface alpha value. Default: 1.0

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap. Default: viridis

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

elevation: Height factor for elevation of 2D plot. Default: 0.0

show: Show plot immediately. Default: true

legend_location: Location of legend. Default: upper right

outline: Plot outline of domain. Default: true

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

label: Label of plot. Default: f

interior: Plot interior of grid. Default: true

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

title: Plot title. Default:

resolution: Plot xy resolution. Default: (500, 500)

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ExtendableGrids.plot — Method

plot(grid; Plotter, kwargs...)

Plot grid without predefined context.

Keyword arguments:

alpha: Surface alpha value. Default: 1.0

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap. Default: viridis

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

elevation: Height factor for elevation of 2D plot. Default: 0.0

show: Show plot immediately. Default: true

legend_location: Location of legend. Default: upper right

outline: Plot outline of domain. Default: true

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

label: Label of plot. Default: f

interior: Plot interior of grid. Default: true

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

title: Plot title. Default:

resolution: Plot xy resolution. Default: (500, 500)

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ExtendableGrids.plottertype — Method

plottertype(p)

Return the type of a plotter.

ExtendableGrids.plottertype — Method

plottertype(Plotter)

Heuristically detect type of plotter, returns the corresponding abstract type fro plotting.

ExtendableGrids.tridata — Method

tridata(grid::Any) -> Tuple{Any,Any,Any}

Return tridata to be splatted to PyPlot calls