Simplex grid constructors

Constructors for simplex grids

API

Base.write — Method

write(fname::String, g::ExtendableGrid; format)

Write grid to file. Currently for pdelib sg format only

ExtendableGrids.geomspace — Method

geomspace(a::Any, b::Any, ha::Any, hb::Any; tol) -> Any

(Try to) create a subdivision of interval (a,b) stored in the returned array X such that

X[1]==a, X[end]==b

(X[2]-X[1])<=ha+tol*(b-a)

(X[end]-X[end-1])<=hb+tol*(b-a)

There is a number q such that X[i+1]-X[i] == q*(X[i]-X[i-1])

X is the array with the minimal possible number of points with the above property

Caveat: the algorithm behind this is well tested but unproven.

Returns an Array containing the points of the subdivision.

ExtendableGrids.glue — Method

glue(a::Array{Tv,1}, b::Array{Tv,1}; tol) -> Array{_A,1} where _A

Glue together two vectors a and b resulting in a vector c. They last element of a shall be equal (up to tol) to the first element of b. The result fulfills length(c)=length(a)+length(b)-1

ExtendableGrids.simplexgrid — Method

simplexgrid(file::Any; format) -> ExtendableGrid{_A,Int32} where _A

Read grid from file. Currently for pdelib sg format only

ExtendableGrids.simplexgrid — Method

Constructor for 1D grid.

Construct 1D grid from an array of node cordinates. It creates two boundary regions with index 1 at the left end and index 2 at the right end.

Primal grid holding unknowns: marked by o, dual grid marking control volumes: marked by |.

 o-----o-----o-----o-----o-----o-----o-----o-----o
 |--|-----|-----|-----|-----|-----|-----|-----|--|

ExtendableGrids.simplexgrid — Method

simplexgrid(file::Tc; format) -> ExtendableGrid{_A,Int32} where _A

Constructor for 3D grid from coordinate arrays. Boundary region numbers:

location	number

south	1

east	2

north	3

west	4

bottom	5

top	6

ExtendableGrids.simplexgrid — Method

simplexgrid(file::Tc; format) -> ExtendableGrid{_A,Int32} where _A

Constructor for 2D grid from coordinate arrays. Boundary region numbers count counterclockwise:

location	number

south	1

east	2

north	3

west	4

ExtendableGrids.simplexgrid — Method

function simplexgrid(coord::Array{Tc,2},
                     cellnodes::Array{Ti,2},
                     cellregions::Array{Ti,1},
                     bfacenodes::Array{Ti,2},
                     bfaceregions::Array{Ti,1}
                     ) where {Tc,Ti}

Create simplex grid from five arrays.