Type hierarchy

The package defines a hierarchy of abstract types to handle grid compoments. The full tree is listed here:

AbstractExtendableGridApexType
├─ AbstractCoordinateSystem
│  ├─ Cartesian1D
│  ├─ Cartesian2D
│  ├─ Cartesian3D
│  ├─ Cylindrical2D
│  ├─ Cylindrical3D
│  ├─ Polar1D
│  ├─ Polar2D
│  ├─ Spherical1D
│  └─ Spherical3D
├─ AbstractElementGeometry
│  ├─ AbstractElementGeometry0D
│  │  └─ Vertex0D
│  ├─ AbstractElementGeometry1D
│  │  └─ Edge1D
│  ├─ AbstractElementGeometry2D
│  │  ├─ Circle2D
│  │  └─ Polygon2D
│  │     ├─ Hexagon2D
│  │     ├─ Pentagon2D
│  │     ├─ Quadrilateral2D
│  │     │  └─ Parallelogram2D
│  │     └─ Triangle2D
│  ├─ AbstractElementGeometry3D
│  │  ├─ Polyhedron3D
│  │  │  ├─ Hexahedron3D
│  │  │  │  └─ Parallelepiped3D
│  │  │  ├─ Prism3D
│  │  │  │  └─ TrianglePrism3D
│  │  │  └─ Tetrahedron3D
│  │  └─ Sphere3D
│  └─ AbstractElementGeometry4D
│     ├─ HyperCube4D
│     └─ Polychoron4D
└─ AbstractGridComponent
   ├─ AbstractElementGeometries
   │  ├─ BFaceGeometries
   │  └─ CellGeometries
   ├─ AbstractElementRegions
   │  ├─ BFaceRegions
   │  └─ CellRegions
   ├─ AbstractGridAdjacency
   │  ├─ BFaceNodes
   │  ├─ CellEdges
   │  ├─ CellNodes
   │  ├─ EdgeCells
   │  └─ EdgeNodes
   ├─ AbstractGridFloatArray1D
   │  ├─ ZCoordinates
   │  ├─ XCoordinates
   │  └─ YCoordinates
   ├─ AbstractGridFloatArray2D
   │  └─ Coordinates
   ├─ AbstractGridIntegerArray1D
   │  └─ NodeInParent
   ├─ AbstractGridIntegerArray2D
   ├─ CoordinateSystem
   ├─ AbstractGridFloatConstant
   ├─ AbstractGridIntegerConstant
   │  ├─ NumBFaceRegions
   │  └─ NumCellRegions
   └─ ParentGrid

API

ExtendableGrids.AbstractExtendableGridApexType — Type

abstract type AbstractExtendableGridApexType

Apex type of all abstract types in this hierarchy.

AbstractTrees.children — Method

children(T::Type) -> Union{Array{Type,1}, Array{Any,1}}

Define children for types.

ExtendableGrids.typehierarchy — Method

typehierarchy()

Print complete type hierachy for ExtendableGrids