Visualization

Provide visualization facilities for grids and functions on grids for various Julia graphics APIs.

In the moment, the following features are available:

Plotting simplex grids with cell and boundary regions

Plotting piecewise linear functions on simplex grids

The basic structure is to create a GridVisiualizer p which allows to have a grid of subplots. The visualize! methods then plot into a subplot p[i,j]. Creation of a GridVisiualizer takes a Plotter keyword argument, which allow to specify plotting module. Supported are:

From the REPL: GLMakie, PyPlot, VTKView (linux only), Plots (no 3D, no unstructured)

From Pluto notebooks: GLMakie PyPlot, Plots, WGLMakie (experimental), MeshCat (experimental)

Note that functionality here mostly will be added on necessity. Eventually, this will be factored out to a standalone package.

API

ExtendableGrids.GridVisualize.SubVis — Type

A SubVis is just a dictionary which contains plotting information, including type of the plotter and its position in the plot.

ExtendableGrids.GridVisualize.GridVisualizer — Type

struct GridVisualizer

Context type for plots.

ExtendableGrids.GridVisualize.GridVisualizer — Method

GridVisualizer(; Plotter, kwargs...)

Create a plot context.

Plotter: defaults to nothing and can be PyPlot, Plots, VTKView, Makie. This pattern to pass the backend as a module to a plot function allows to circumvent to create heavy default package dependencies.

Depending on the layout keyword argument, a 2D grid of subplots is created. Further visualize! commands then plot into one of these subplots:

p=GridVisualizer(Plotter=PyPlot, layout=(2,2)
visualize!(p[1,2], ...)
````

A `plot`  command just implicitely creates a plot context:

julia visualize(..., Plotter=PyPlot)

is equivalent to

julia p=GridVisualizer(Plotter=PyPlot, layout=(1,1) visualize!(p,...) ```

Please not that the return values of all plot commands are specific to the Plotter.

Depending on the backend, interactive mode switch between "gallery view" showing all plots at onece and "focused view" showing only one plot is possible.

Keyword arguments:

outline: Plot outline of domain. Default: true

alpha: Surface alpha value. Default: 1.0

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap (any from ColorSchemes.jl). Default: viridis

label: Label of plot. Default:

cellwise: Cellwise 1D plot, can be slow). Default: false

interior: Plot interior of grid. Default: true

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

fast: Fast updates (with Makie, restricted functionality). Default: true

fignumber: Figure number (PyPlot). Default: 1

title: Plot title. Default:

framepos: Subplot position in frame (VTKView). Default: 1

zlimits: z limits. Default: (1, -1)

flimits: function limits. Default: (1, -1)

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

reveal: Show plot immediately (same as :show). Default: false

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

resolution: Plot xy resolution. Default: (500, 500)

elevation: Height factor for elevation of 2D plot. Default: 0.0

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

show: Show plot immediately. Default: false

xlimits: x limits. Default: (1, -1)

legend_location: Location of legend. Default: upper right

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ylimits: y limits. Default: (1, -1)

ExtendableGrids.GridVisualize.MakieType — Type

abstract type MakieType

Abstract type for dispatching on plotter

ExtendableGrids.GridVisualize.MeshCatType — Type

abstract type MeshCatType

Abstract type for dispatching on plotter

ExtendableGrids.GridVisualize.PlotsType — Type

abstract type PlotsType

Abstract type for dispatching on plotter

ExtendableGrids.GridVisualize.PyPlotType — Type

abstract type PyPlotType

Abstract type for dispatching on plotter

ExtendableGrids.GridVisualize.VTKViewType — Type

abstract type VTKViewType

Abstract type for dispatching on plotter

Base.getindex — Method

getindex(p, i, j)

Return a SubVis

Base.size — Method

size(p)

Return the layout of a GridVisualizer

ExtendableGrids.GridVisualize.ismakie — Method

ismakie(Plotter)

Heuristically check if Plotter is Makie/WGLMakie

ExtendableGrids.GridVisualize.ismeshcat — Method

ismeshcat(Plotter)

Heuristically check if Plotter is MeshCat

ExtendableGrids.GridVisualize.isplots — Method

isplots(Plotter)

Heuristically check if Plotter is Plots

ExtendableGrids.GridVisualize.ispyplot — Method

ispyplot(Plotter)

Heuristically check if Plotter is PyPlot

ExtendableGrids.GridVisualize.isvtkview — Method

isvtkview(Plotter)

Heuristically check if Plotter is VTKView

ExtendableGrids.GridVisualize.plottertype — Method

plottertype(p)

Return the type of a plotter.

ExtendableGrids.GridVisualize.plottertype — Method

plottertype(Plotter)

Heuristically detect type of plotter, returns the corresponding abstract type fro plotting.

ExtendableGrids.GridVisualize.reveal — Method

reveal(p)

Finish and show plot. Same as setting :reveal=true or :show=true in last visualize statment for a context.

ExtendableGrids.GridVisualize.save — Method

save(fname, p)

Save figure to disk

ExtendableGrids.GridVisualize.visualize! — Method

visualize!(p, grid; kwargs...)

Plot grid.

Keyword arguments:

outline: Plot outline of domain. Default: true

alpha: Surface alpha value. Default: 1.0

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap (any from ColorSchemes.jl). Default: viridis

label: Label of plot. Default:

cellwise: Cellwise 1D plot, can be slow). Default: false

interior: Plot interior of grid. Default: true

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

fast: Fast updates (with Makie, restricted functionality). Default: true

fignumber: Figure number (PyPlot). Default: 1

title: Plot title. Default:

framepos: Subplot position in frame (VTKView). Default: 1

zlimits: z limits. Default: (1, -1)

flimits: function limits. Default: (1, -1)

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

reveal: Show plot immediately (same as :show). Default: false

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

resolution: Plot xy resolution. Default: (500, 500)

elevation: Height factor for elevation of 2D plot. Default: 0.0

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

show: Show plot immediately. Default: false

xlimits: x limits. Default: (1, -1)

legend_location: Location of legend. Default: upper right

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ylimits: y limits. Default: (1, -1)

ExtendableGrids.GridVisualize.visualize! — Method

visualize!(ctx, grid, func; kwargs...)

Plot vector on grid as P1 FEM function.

Keyword arguments

outline: Plot outline of domain. Default: true

alpha: Surface alpha value. Default: 1.0

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap (any from ColorSchemes.jl). Default: viridis

label: Label of plot. Default:

cellwise: Cellwise 1D plot, can be slow). Default: false

interior: Plot interior of grid. Default: true

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

fast: Fast updates (with Makie, restricted functionality). Default: true

fignumber: Figure number (PyPlot). Default: 1

title: Plot title. Default:

framepos: Subplot position in frame (VTKView). Default: 1

zlimits: z limits. Default: (1, -1)

flimits: function limits. Default: (1, -1)

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

reveal: Show plot immediately (same as :show). Default: false

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

resolution: Plot xy resolution. Default: (500, 500)

elevation: Height factor for elevation of 2D plot. Default: 0.0

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

show: Show plot immediately. Default: false

xlimits: x limits. Default: (1, -1)

legend_location: Location of legend. Default: upper right

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ylimits: y limits. Default: (1, -1)

ExtendableGrids.GridVisualize.visualize! — Method

visualize!(ctx, grid; kwargs...)

Plot grid.

Keyword arguments:

outline: Plot outline of domain. Default: true

alpha: Surface alpha value. Default: 1.0

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap (any from ColorSchemes.jl). Default: viridis

label: Label of plot. Default:

cellwise: Cellwise 1D plot, can be slow). Default: false

interior: Plot interior of grid. Default: true

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

fast: Fast updates (with Makie, restricted functionality). Default: true

fignumber: Figure number (PyPlot). Default: 1

title: Plot title. Default:

framepos: Subplot position in frame (VTKView). Default: 1

zlimits: z limits. Default: (1, -1)

flimits: function limits. Default: (1, -1)

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

reveal: Show plot immediately (same as :show). Default: false

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

resolution: Plot xy resolution. Default: (500, 500)

elevation: Height factor for elevation of 2D plot. Default: 0.0

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

show: Show plot immediately. Default: false

xlimits: x limits. Default: (1, -1)

legend_location: Location of legend. Default: upper right

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ylimits: y limits. Default: (1, -1)

ExtendableGrids.GridVisualize.visualize — Method

visualize(grid, func; Plotter, kwargs...)

Plot vector on grid without predefined context as P1 FEM function.

Keyword arguments:

outline: Plot outline of domain. Default: true

alpha: Surface alpha value. Default: 1.0

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap (any from ColorSchemes.jl). Default: viridis

label: Label of plot. Default:

cellwise: Cellwise 1D plot, can be slow). Default: false

interior: Plot interior of grid. Default: true

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

fast: Fast updates (with Makie, restricted functionality). Default: true

fignumber: Figure number (PyPlot). Default: 1

title: Plot title. Default:

framepos: Subplot position in frame (VTKView). Default: 1

zlimits: z limits. Default: (1, -1)

flimits: function limits. Default: (1, -1)

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

reveal: Show plot immediately (same as :show). Default: false

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

resolution: Plot xy resolution. Default: (500, 500)

elevation: Height factor for elevation of 2D plot. Default: 0.0

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

show: Show plot immediately. Default: false

xlimits: x limits. Default: (1, -1)

legend_location: Location of legend. Default: upper right

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ylimits: y limits. Default: (1, -1)

ExtendableGrids.GridVisualize.visualize — Method

visualize(grid; Plotter, kwargs...)

Plot grid without predefined context.

Keyword arguments:

outline: Plot outline of domain. Default: true

alpha: Surface alpha value. Default: 1.0

colorbar: Show colorbar in plots. Default: true

subplot: Actual subplot. Default: (1, 1)

zplane: zplane for 3D visualization. Default: 1.7976931348623157e308

colormap: Contour plot colormap (any from ColorSchemes.jl). Default: viridis

label: Label of plot. Default:

cellwise: Cellwise 1D plot, can be slow). Default: false

interior: Plot interior of grid. Default: true

flevel: isolevel for 3D visualization. Default: 1.7976931348623157e308

elev: Elevation angle for 3D visualization (in degrees). Default: 30

azim: Azimuth angle for 3D visualization (in degrees). Default: -60

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

fast: Fast updates (with Makie, restricted functionality). Default: true

fignumber: Figure number (PyPlot). Default: 1

title: Plot title. Default:

framepos: Subplot position in frame (VTKView). Default: 1

zlimits: z limits. Default: (1, -1)

flimits: function limits. Default: (1, -1)

yplane: yplane for 3D visualization. Default: 1.7976931348623157e308

aspect: Aspect ratio modification. Default: 1.0

reveal: Show plot immediately (same as :show). Default: false

clear: Clear plot before new plot.. Default: true

legend: Add legend to plot. Default: true

resolution: Plot xy resolution. Default: (500, 500)

elevation: Height factor for elevation of 2D plot. Default: 0.0

layout: Layout of plots. Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

color: Color of lines on plot. Default: (0, 0, 0)

show: Show plot immediately. Default: false

xlimits: x limits. Default: (1, -1)

legend_location: Location of legend. Default: upper right

edges: Plot grid edges when plotting grid. Default: true

xplane: xplane for 3D visualization. Default: 1.7976931348623157e308

ylimits: y limits. Default: (1, -1)