Adjacency

This handles adjacency matrices between entities of polyhedral complexes, e.g. nodes, cells, edges etc.

An adjacency is described by an Adjacency matrix, which is a sparse matrix whose entries a 0 or 1. While such a matrix always can be stored as a SparseMatrixCSC, in general this would be a waste of storage.

For the general case, it is sufficient to only store the column start indieces and the column entries (row numbers), and to implicitely assume that nonzero entries are 1. This kind of storage is realised in a VariableTargetAdjacency.

In many cases, this can be compressed even more, if each column has the same length. In that case, a Matrix is sufficient to store the data. This is the usual base for implementing FEM/FVM assembly, and the interface for the general case should be similar.

From these ideas we develop the following interface for an adjacency a.

In order to avoid name confusion, we introduce the following notation which should be consistent with the use in assembly loops.

source: source of adjacency link target: target of adjacency link

E.g. the cell-node adjacency for FEM assembly links a number of cells with a collection of nodes. The cells are the sources, and the targets are the nodes.

getindex(a,i,isource) aka a[i,isource]: return i-th target of source j numsources(a): overall number of sources, e.g. number of cells numtargets(a): overall number of targets numtargets(a,isource): number of targets for source given by isource numlinks(a): number of links aka nonzero entries of adjacency matrix show(a): print stuff

Further API ideas:

Convert between Matrix and Variable target stuff using 0 entries as "padding"

API

ExtendableGrids.Adjacency — Constant

Adjacency type as union of FixedTargetAdjacency and VariableTargetAdjacency

ExtendableGrids.Adjacency — Method

Constructors for Adjacency

ExtendableGrids.FixedTargetAdjacency — Type

mutable struct Array{T, 2} <: DenseArray{T,2}

Use Matrix to store fixed target adjacency

ExtendableGrids.VariableTargetAdjacency — Type

struct VariableTargetAdjacency{T}

Adjacency struct. Essentially, this is the sparsity pattern of a matrix whose nonzero elements all have the same value in the CSC format.

ExtendableGrids.VariableTargetAdjacency — Method

VariableTargetAdjacency() -> VariableTargetAdjacency{Int64}

Create an empty VariableTargetAdjacency with default type

ExtendableGrids.VariableTargetAdjacency — Method

Create a VariableTargetAdjacency from Matrix

ExtendableGrids.VariableTargetAdjacency — Method

Create an empty VariableTargetAdjacency

Base.:== — Method

==(a, b)

Comparison of two adjacencies

Base.append! — Method

append!(adj::VariableTargetAdjacency, column::Any) -> Array{T,1} where T

Append a column to adjacency.

Base.getindex — Method

getindex(adj::VariableTargetAdjacency, i::Any, isource::Any) -> Any

Access adjacency as if it is a 2D Array

Base.show — Method

show(io::IO, adj::VariableTargetAdjacency)

Show adjacency (in trasposed form; preliminary)

ExtendableGrids.atranspose — Method

Transpose adjacency

ExtendableGrids.max_num_targets_per_source — Method

max_num_targets_per_source(adj::Array{T,2} where T) -> Int64

Maximum number of targets per source

ExtendableGrids.max_num_targets_per_source — Method

max_num_targets_per_source(adj::VariableTargetAdjacency) -> Any

Maximum number of targets per source

ExtendableGrids.num_links — Method

num_links(adj::Array{T,2} where T) -> Int64

Number of entries

ExtendableGrids.num_links — Method

num_links(adj::VariableTargetAdjacency) -> Int64

Number of links

ExtendableGrids.num_sources — Method

num_sources(adj::Array{T,2} where T) -> Int64

Number of sources in adjacency

ExtendableGrids.num_sources — Method

num_sources(adj::VariableTargetAdjacency) -> Int64

Number of sources in adjacency

ExtendableGrids.num_targets — Method

num_targets(adj::Array{T,2} where T, isource::Any) -> Int64

Number of targets per source if adjacency is a matrix

ExtendableGrids.num_targets — Method

num_targets(adj::Array{T,2} where T) -> Any

Overall number of targets

ExtendableGrids.num_targets — Method

num_targets(adj::VariableTargetAdjacency, isource::Any) -> Any

Number of targets for given source

ExtendableGrids.num_targets — Method

num_targets(adj::VariableTargetAdjacency) -> Any

Number of targeta