Plotting

Provide plotting facilities for various Julia graphics APIs. A graphics API (Plots, PyPlot, VTKView) needs to be passed to plot in the Plotter keyword parameter

API

ExtendableGrids.isplots — Method

isplots(Plotter::Any) -> Bool

Check if Plotter is Plots

ExtendableGrids.ispyplot — Method

ispyplot(Plotter::Any) -> Bool

Check if Plotter is PyPlot

ExtendableGrids.isvtkview — Method

isvtkview(Plotter::Any) -> Bool

Check if Plotter is VTKView

ExtendableGrids.plot — Method

plot(grid::ExtendableGrid, U::AbstractArray{T,1} where T; Plotter, color, cmap, label, levels, aspect, clear, show, p)

Plot lowest order continuous finite element function on grid defined by values on the grid nodes.

Keyword arguments:

Plotter: defaults to nothing and can be PyPLot, Plots, VTKView.

color: color of plot on 1D grid

cmap: color map for heatmap plot

label: label of plot

levels: number of isolevels

aspect: aspect ratio of plot

clear: if true (default) clear plot before plotting

show: if true (default) show plot

ExtendableGrids.plot — Method

plot(grid::ExtendableGrid; Plotter)

Plot grid.

Plotter defaults to nothing and can be PyPLot, Plots, VTKView.

ExtendableGrids.frgb — Method

frgb(Plotter::Any, i::Any, max::Any; pastel) -> Any

Color scale for grid colors.

ExtendableGrids.tridata — Method

tridata(grid::Any) -> Tuple{Any,Any,Any}

Return tridata to be splatted to PyPlot calls