Plotting

Provide plotting facilities for various Julia graphics APIs.

API

ExtendableGrids.MakieType — Type

abstract type MakieType

Abstract type for dispatching on plotter

ExtendableGrids.PlotsType — Type

abstract type PlotsType

Abstract type for dispatching on plotter

ExtendableGrids.PlotterContext — Type

struct PlotterContext <: AbstractDict{Any,Any}

Context type for plots. Wrapper around dict.

ExtendableGrids.PlotterContext — Method

PlotterContext(Plotter; kwargs...)

Initalize plotter context

Plotter: defaults to nothing and can be PyPlot, Plots, VTKView, Makie.

Keyword arguments:

colorbar: Show colorbar in plots. Default: true

aspect: Aspect ration modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

subplot: Subplot number for multiple plots. Default: 111

colormap: Contour plot colormap. Default: hot

legend: Add legend to plot. Default: true

label: Label of plot. Default: f

elevation_factor: Height factor for elevation. Default: 1.0

elevation: Show elevation plot (2D function graph). Default: false

layout: Layout of multiple plots (experimental). Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

resolution: Plot resolution for pairs. Default: (300, 300)

color: Color of lines on plot. Default: (0, 0, 0)

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

show: Show plots immediately. Default: true

legend_location: Location of legend. Default: upper right

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

ExtendableGrids.PyPlotType — Type

abstract type PyPlotType

Abstract type for dispatching on plotter

ExtendableGrids.VTKViewType — Type

abstract type VTKViewType

Abstract type for dispatching on plotter

ExtendableGrids.frgb — Method

frgb(Plotter, i, max; pastel)

Color scale for grid colors.

ExtendableGrids.ismakie — Method

ismakie(Plotter)

Heuristically check if Plotter is Makie/WGLMakie

ExtendableGrids.isplots — Method

isplots(Plotter)

Heuristically check if Plotter is Plots

ExtendableGrids.ispyplot — Method

ispyplot(Plotter)

Heuristically check if Plotter is PyPlot

ExtendableGrids.isvtkview — Method

isvtkview(Plotter)

Heuristically check if Plotter is VTKView

ExtendableGrids.plot! — Method

plot!(ctx, grid, func; kwargs...)

Plot vector on grid as P1 FEM function.

Keyword arguments:

colorbar: Show colorbar in plots. Default: true

aspect: Aspect ration modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

subplot: Subplot number for multiple plots. Default: 111

colormap: Contour plot colormap. Default: hot

legend: Add legend to plot. Default: true

label: Label of plot. Default: f

elevation_factor: Height factor for elevation. Default: 1.0

elevation: Show elevation plot (2D function graph). Default: false

layout: Layout of multiple plots (experimental). Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

resolution: Plot resolution for pairs. Default: (300, 300)

color: Color of lines on plot. Default: (0, 0, 0)

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

show: Show plots immediately. Default: true

legend_location: Location of legend. Default: upper right

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

ExtendableGrids.plot! — Method

plot!(ctx, grid; kwargs...)

Plot grid.

Keyword arguments:

colorbar: Show colorbar in plots. Default: true

aspect: Aspect ration modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

subplot: Subplot number for multiple plots. Default: 111

colormap: Contour plot colormap. Default: hot

legend: Add legend to plot. Default: true

label: Label of plot. Default: f

elevation_factor: Height factor for elevation. Default: 1.0

elevation: Show elevation plot (2D function graph). Default: false

layout: Layout of multiple plots (experimental). Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

resolution: Plot resolution for pairs. Default: (300, 300)

color: Color of lines on plot. Default: (0, 0, 0)

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

show: Show plots immediately. Default: true

legend_location: Location of legend. Default: upper right

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

ExtendableGrids.plot — Method

plot(grid, func; Plotter, kwargs...)

Plot vector on grid without predefined context as P1 FEM function.

Keyword arguments:

colorbar: Show colorbar in plots. Default: true

aspect: Aspect ration modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

subplot: Subplot number for multiple plots. Default: 111

colormap: Contour plot colormap. Default: hot

legend: Add legend to plot. Default: true

label: Label of plot. Default: f

elevation_factor: Height factor for elevation. Default: 1.0

elevation: Show elevation plot (2D function graph). Default: false

layout: Layout of multiple plots (experimental). Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

resolution: Plot resolution for pairs. Default: (300, 300)

color: Color of lines on plot. Default: (0, 0, 0)

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

show: Show plots immediately. Default: true

legend_location: Location of legend. Default: upper right

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

ExtendableGrids.plot — Method

plot(grid; Plotter, kwargs...)

Plot grid without predefined context.

Keyword arguments:

colorbar: Show colorbar in plots. Default: true

aspect: Aspect ration modification. Default: 1.0

clear: Clear plot before new plot.. Default: true

subplot: Subplot number for multiple plots. Default: 111

colormap: Contour plot colormap. Default: hot

legend: Add legend to plot. Default: true

label: Label of plot. Default: f

elevation_factor: Height factor for elevation. Default: 1.0

elevation: Show elevation plot (2D function graph). Default: false

layout: Layout of multiple plots (experimental). Default: (1, 1)

axisgrid: Show background grid in plots. Default: true

resolution: Plot resolution for pairs. Default: (300, 300)

color: Color of lines on plot. Default: (0, 0, 0)

colorlevels: Number of color levels for contour plot. Default: 51

isolines: Number of isolines in contour plot. Default: 11

show: Show plots immediately. Default: true

legend_location: Location of legend. Default: upper right

fignumber: Figure number (PyPlot). Default: 1

framepos: Subplot position in frame (VTKView). Default: 1

ExtendableGrids.plottertype — Method

plottertype(Plotter)

Heuristically detect type of plotter, returns the corresponding abstract type fro plotting.

ExtendableGrids.tridata — Method

tridata(grid::Any) -> Tuple{Any,Any,Any}

Return tridata to be splatted to PyPlot calls