Region editing

Tools for editing grid region numbers

API

ExtendableGrids.bfacemask! — Method

bfacemask!(grid::ExtendableGrid, maskmin::AbstractArray, maskmax::AbstractArray, ireg::Int64; tol) -> ExtendableGrid

Edit region numbers of grid boundary facets via rectangular mask. For 1D grids, inner boundaries can be added by this method.

ExtendableGrids.cellmask! — Method

cellmask!(grid::ExtendableGrid, maskmin::AbstractArray, maskmax::AbstractArray, ireg::Int64; tol) -> Any

Edit region numbers of grid cells via rectangular mask.