Subgrid

Subgrids of an ExtendableGrid are again of the same type ExtendableGrid and unse the typed Dict mechanism to store linkage to the parent grid.

grid=simplexgrid([1,2,3], [4,5,6])
sub=subgrid(grid,[2],boundary=true, transform=(a,b) -> (a[1]=10*b[2]))
println(keys(sub))
println(sub[Coordinates])

Type{var"#s16"} where var"#s16"<:AbstractGridComponent[CellGeometries, Coordinates, ExtendableGrids.NodeInParent, ExtendableGrids.ParentGrid, CellNodes, CellRegions, NumBFaceRegions]
[40 50 60]

Given a vector on the parent grid, one can create a view of this vecotor on the subgrid:

grid=simplexgrid([1,2,3], [4,5,6])
sub=subgrid(grid,[2],boundary=true, transform=(a,b) -> (a[1]=10*b[2]))
v=[i for i=1:num_nodes(grid)]
subv=view(v,sub)
println(subv)

[3, 6, 9]

API

ExtendableGrids.NodeInParent — Type

abstract type NodeInParent <: AbstractGridIntegerArray1D

Grid component key type for storing node in parent array

ExtendableGrids.ParentGrid — Type

abstract type ParentGrid <: AbstractGridComponent

Grid component key type for storing parent grid

ExtendableGrids.SubgridVectorView — Type

struct SubgridVectorView{Tv, Ti} <: AbstractArray{Tv,1}

Vector view on subgrid

sysarray::AbstractArray{Tv,1} where Tv

node_in_parent::Array{Ti,1} where Ti

Base.getindex — Method

getindex(aview::ExtendableGrids.SubgridVectorView, inode::Integer) -> Any

Accessor method for subgrid vector view.

Base.setindex! — Method

setindex!(aview::ExtendableGrids.SubgridVectorView, v::Any, inode::Integer) -> ExtendableGrids.SubgridVectorView

Accessor method for subgrid vector view.

Base.size — Method

size(a::ExtendableGrids.SubgridVectorView) -> Tuple{Int64}

Return size of vector view.

Base.view — Method

view(a::AbstractArray{T,1} where T, subgrid::ExtendableGrid) -> ExtendableGrids.SubgridVectorView{_A,_B} where _B where _A

Create a view of the vector on a subgrid.

ExtendableGrids._copytransform! — Method

_copytransform!(a::AbstractArray, b::AbstractArray)

Default transform for subgrid creation

ExtendableGrids.subgrid — Method

subgrid(parent::Any, subregions::AbstractArray; transform, boundary, project) -> ExtendableGrid{_A,_B} where _B where _A

Create subgrid from list of regions.

parent: parent grid

subregions: Array of subregions

transform (kw parameter): transformation function between grid and subgrid coordinates acting on one point. Default: copytransform

boundary: if true, create codimension 1 subgrid from boundary region.

A subgrid is of type ExtendableGrid and stores two additional components: ParentGrid and NodeInParent